Fix `lerpLog` casting to `int` instead of `float`

+++ src/main/java/com/kyanite/deeperdarker/util/DDUtil.java
@@ -3,6 +3,6 @@ package com.kyanite.deeperdarker.util;
 public class DDUtil {
     public static float lerpLog(float t, float a, float b) {
         // https://www.cmu.edu/biolphys/deserno/pdf/log_interpol.pdf
-        return (int) (Math.pow(a, t) * Math.pow(b, 1.0f - t));
+        return (float) (Math.pow(a, t) * Math.pow(b, 1.0f - t));
     }
 }
